+++ upstash_workflow/serve/options.py
@@ -43,7 +43,9 @@ class ServeOptions(Generic[TInitialPayload, TResponse]):
     retries: int
     url: Optional[str]
     failure_function: Optional[
-        Callable[[WorkflowContext[TInitialPayload], int, str, Dict[str, str]], Any]
+        Callable[
+            [WorkflowContext[TInitialPayload], int, Optional[str], Dict[str, str]], Any
+        ]
     ]
     failure_url: Optional[str]
 
@@ -66,7 +68,7 @@ def _process_options(
     retries: Optional[int] = DEFAULT_RETRIES,
     url: Optional[str] = None,
     failure_function: Optional[
-        Callable[[WorkflowContext, int, str, Dict[str, str]], Any]
+        Callable[[WorkflowContext, int, Optional[str], Dict[str, str]], Any]
     ] = None,
     failure_url: Optional[str] = None,
 ) -> ServeBaseOptions[TInitialPayload, TResponse]:
@@ -83,7 +85,7 @@ def _process_options(
     - retries: DEFAULT_RETRIES
     - url: None
     """
-    environment = env if env is not None else dict(os.environ)
+    environment: Dict[str, Optional[str]] = env if env is not None else dict(os.environ)
 
     receiver_environment_variables_set = bool(
         environment.get("QSTASH_CURRENT_SIGNING_KEY")